relocate coverage lines and make ItemIds internal
