+++ ddtrace/ext/test_visibility/api.py
@@ -24,6 +24,9 @@ from typing import Optional
 from typing import Type
 
 from ddtrace.ext.test import Status as _TestStatus
+from ddtrace.ext.test_visibility._item_ids import TestId
+from ddtrace.ext.test_visibility._item_ids import TestModuleId
+from ddtrace.ext.test_visibility._item_ids import TestSuiteId
 from ddtrace.ext.test_visibility._test_visibility_base import TestSessionId
 from ddtrace.ext.test_visibility._test_visibility_base import TestSourceFileInfoBase
 from ddtrace.ext.test_visibility._test_visibility_base import TestVisibilityItemId
@@ -35,9 +38,6 @@ from ddtrace.ext.test_visibility._utils import _get_item_tag
 from ddtrace.ext.test_visibility._utils import _is_item_finished
 from ddtrace.ext.test_visibility._utils import _set_item_tag
 from ddtrace.ext.test_visibility._utils import _set_item_tags
-from ddtrace.ext.test_visibility.item_ids import TestId
-from ddtrace.ext.test_visibility.item_ids import TestModuleId
-from ddtrace.ext.test_visibility.item_ids import TestSuiteId
 from ddtrace.internal import core
 from ddtrace.internal.logger import get_logger as _get_logger
 import ddtrace.tracing  # noqa: F401 -- this triggers the registration of trace handlers
