+++ ddtrace/internal/ci_visibility/api/_coverage_data.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from typing import Dict
 from typing import List
 
-from ddtrace.ext.test_visibility.coverage_lines import CoverageLines
+from ddtrace.internal.test_visibility.coverage_lines import CoverageLines
 
 
 try:
