+++ tests/ci_visibility/api/fake_runner_mix_fail_itr_test_level.py
@@ -22,8 +22,8 @@ import sys
 from unittest import mock
 
 from ddtrace.ext.test_visibility import api as ext_api
-from ddtrace.ext.test_visibility.coverage_lines import CoverageLines
 from ddtrace.internal.test_visibility import api
+from ddtrace.internal.test_visibility.coverage_lines import CoverageLines
 
 
 def _make_excinfo():
