+++ tests/ci_visibility/test_ci_visibility.py
@@ -19,7 +19,7 @@ from ddtrace.constants import AUTO_KEEP
 from ddtrace.ext import ci
 from ddtrace.ext.git import _build_git_packfiles_with_details
 from ddtrace.ext.git import _GitSubprocessDetails
-import ddtrace.ext.test_visibility.item_ids
+import ddtrace.ext.test_visibility._item_ids
 from ddtrace.internal.ci_visibility import CIVisibility
 from ddtrace.internal.ci_visibility.constants import REQUESTS_MODE
 from ddtrace.internal.ci_visibility.constants import SUITE
